requirements.txt

absl-py
asgiref
astunparse
cachetools
certifi
charset-normalizer
click
cloudinary
cmake
colorama
cycler
distlib
Django
django-cloudinary-storage
django-cors-headers
djangorestframework

filelock
Flask
flatbuffers
fonttools
gast
google-auth
google-auth-oauthlib
google-pasta
grpcio
h5py
idna
ijson
imageio
imutils
itsdangerous
Jinja2
Keras-Preprocessing
kiwisolver
libclang
Markdown
MarkupSafe
matplotlib
networkx
oauthlib
opencv-python
opt-einsum
packaging
Pillow
platformdirs
playsound
protobuf
pyasn1
pyasn1-modules
pygame
Pygments
pyparsing
PySimpleGUI
pysnow
python-dateutil
python-dotenv
python-magic
python-settings
pytz
PyWavelets
PyYAML
requests
requests-oauthlib
rsa
scikit-image
scipy
six
sqlparse
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorflow
tensorflow-io-gcs-filesystem
termcolor
tf-estimator-nightly
tifffile

typing_extensions
tzdata
urllib3
virtualenv
Werkzeug
wrapt
